django/django_uwsgi_nginx_otel/requirements.txt
django
uwsgi
opentelemetry-distro
opentelemetry-exporter-otlp